+++ apps/storybook/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @pikas-ui/storybook
 
+## 1.0.13
+
+### Patch Changes
+
+- Updated dependencies
+  - @pikas-ui/button@1.0.13
+  - @pikas-ui/icons@1.0.13
+  - @pikas-ui/loader@1.0.13
+  - @pikas-ui/separator@1.0.13
+  - @pikas-ui/styles@1.0.13
+  - @pikas-ui/tooltip@1.0.13
+
 ## 1.0.12
 
 ### Patch Changes