+++ packages/button/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @pikas-ui/button
 
+## 1.0.13
+
+### Patch Changes
+
+- Add Tooltip component
+- Updated dependencies
+  - @pikas-ui/icons@1.0.13
+  - @pikas-ui/loader@1.0.13
+  - @pikas-ui/styles@1.0.13
+
 ## 1.0.12
 
 ### Patch Changes