+++ packages/styles/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pikas-ui/styles
 
+## 1.0.13
+
+### Patch Changes
+
+- Add Tooltip component
+
 ## 1.0.12
 
 ### Patch Changes